autorosation de atelier et module

+++ core/app/Http/Controllers/AtelierController.php
@@ -6,14 +6,17 @@ use App\Models\Atelier;
 use App\Http\Requests\StoreAtelierRequest;
 use App\Http\Requests\UpdateAtelierRequest;
 use App\Http\Resources\AtelierResource;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class AtelierController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny');
         return AtelierResource::collection(Atelier::all());
 
     }
@@ -23,6 +26,7 @@ class AtelierController extends Controller
      */
     public function store(StoreAtelierRequest $request)
     {
+        $this->authorize('create');
         $formFields = $request ->validated();
         $atelier = Atelier::create($formFields);
         $response = new AtelierResource($atelier);
@@ -37,6 +41,7 @@ class AtelierController extends Controller
      */
     public function show(Atelier $atelier)
     {
+        $this->authorize('view', $atelier);
         return new AtelierResource($atelier);
     }
 
@@ -45,6 +50,7 @@ class AtelierController extends Controller
      */
     public function update(UpdateAtelierRequest $request, Atelier $atelier)
     {
+        $this->authorize('update', $atelier);
         $formFields =$request->validated();
         $atelier->update($formFields);
         return response()->json([
@@ -58,6 +64,8 @@ class AtelierController extends Controller
      */
     public function destroy(Atelier $atelier)
     {
+        $this->authorize('delete', $atelier);
+        $this->authorize('delete', $atelier);
         $atelier->delete();
         return response()->json([
             'atelier' => $atelier,
