+++ core/app/Http/Controllers/ModuleController.php
@@ -6,14 +6,17 @@ use App\Models\Module;
 use App\Http\Requests\StoreModuleRequest;
 use App\Http\Requests\UpdateModuleRequest;
 use App\Http\Resources\ModuleResource;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ModuleController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny');
         return ModuleResource::collection(Module::all());
     }
 
@@ -22,6 +25,7 @@ class ModuleController extends Controller
      */
     public function store(StoreModuleRequest $request)
     {
+        $this->authorize('create');
         $formFields = $request->validated();
         $response = Module::create($formFields);
         $module = new ModuleResource($response);
@@ -36,6 +40,7 @@ class ModuleController extends Controller
      */
     public function show(Module $module)
     {
+        $this->authorize('view', $module);
         return new ModuleResource($module);
     }
 
@@ -44,6 +49,7 @@ class ModuleController extends Controller
      */
     public function update(UpdateModuleRequest $request, Module $module)
     {
+        $this->authorize('update', $module);
         $formFields =$request->validated();
         $module->update($formFields);
         return response()->json([
@@ -57,6 +63,7 @@ class ModuleController extends Controller
      */
     public function destroy(Module $module)
     {
+        $this->authorize('delete', $module);
         $module->delete();
         return response()->json([
             'module' => $module,
